html/semantics/invokers/invoketarget-on-video-invalid-behavior.tentative.html

<!doctype html>
<meta charset="utf-8" />
<meta name="author" title="Keith Cirkel" href="mailto:wpt@keithcirkel.co.uk" />
<meta name="timeout" content="long" />
<link rel="help" href="https://open-ui.org/components/invokers.explainer/" />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="resources/invoker-utils.js"></script>

<video controls id="invokee" src="/media/movie_5.mp4"></video>
<button id="invokerbutton" invoketarget="invokee"></button>

<script>
  // invalid actions on video
  [
    "foo-bar",
    "showpopover",
    "showmodal",
    "showpicker",
    "open",
    "close",
  ].forEach((action) => {
    promise_test(async function (t) {
      t.add_cleanup(() => invokerbutton.removeAttribute("invokeaction"));
      invokerbutton.setAttribute("invokeaction", action);
      assert_true(invokee.paused);
      assert_false(invokee.muted);
      await clickOn(invokerbutton);
      await waitForRender();
      assert_true(invokee.paused);
      assert_false(invokee.muted);
    }, `invoking (as ${action}) on video does nothing`);
  });
</script>
